# swagger.yaml
openapi: 3.0.0
info:
  title: TicketRush API
  description: Hệ thống phân phối vé sự kiện và quản lý đặt chỗ thời gian thực
  version: 1.0.0
  contact:
    name: Support Team
...
servers:
  - url: http://localhost:3000
    description: Local Development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Sử dụng JWT token từ Authorization header
    
    cookieAuth:
      type: apiKey
      in: cookie
      name: jwt
      description: |
        🍪 HttpOnly Cookie Authentication
        
        ✅ Quá trình:
        1. Gọi /api/auth/login để nhận HttpOnly cookie
        2. Cookie tự động được gửi kèm theo mỗi request (nếu credentials được bật)
        3. Các endpoint sẽ xác thực dựa trên cookie JWT này
        
        📌 Lưu ý:
        - HttpOnly Cookie không thể truy cập từ JavaScript (XSS protection)
        - Browser sẽ tự động gửi cookie giữa các request cùng origin
        - Khi test cross-domain, cần bật CORS credentials
        
        🔒 Security attributes:
        - HttpOnly: true (chống XSS)
        - Secure: true (chỉ qua HTTPS)
        - SameSite: None (cho phép cross-domain)
paths:
...
  /api/Booking/hold:
    post:
      tags: [Booking]
      summary: Giữ ghế để đặt hàng
      description: Giữ ghế trong 1 phút để hoàn tất thanh toán
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - eventId
                - seatIds
              properties:
                eventId:
                  type: string
                seatIds:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Giữ ghế thành công
        '412':
          description: Một số ghế không còn trống
  /api/Booking/checkout:
    post:
      tags: [Booking]
      summary: Hoàn tất thanh toán
      description: Xác nhận mua các ghế đã giữ và tạo đơn hàng
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - eventId
              properties:
                eventId:
                  type: string
      responses:
        '201':
          description: Thanh toán thành công
...
  /api/Booking/return:
    post:
      tags: [Booking]
      summary: Trả ghế và giải phóng slot hàng chờ
      description: |
        Nhả toàn bộ ghế đang ở trạng thái `LOCKED` về `AVAILABLE` và giải phóng
        slot trong Virtual Queue để người tiếp theo được vào.

        ### Khi nào gọi API này?
        Gọi API này khi **người dùng chủ động hủy** quá trình đặt vé, ví dụ:
        - Bấm nút "← Huỷ & Quay Lại Chọn Ghế" trên trang Checkout.
        - Bấm nút "Huỷ" / "Quay lại" ở bất kỳ bước nào trước khi thanh toán xong.

        ### Lưu ý quan trọng cho Frontend
        > ⚠️ **KHÔNG** gọi API này khi điều hướng từ trang EventDetail sang trang Checkout.
        > Khi user bấm "Giữ Chỗ & Thanh Toán", frontend cần dùng một **flag nội bộ**
        > (ví dụ: `isNavigatingToCheckout = true`) để tránh hàm cleanup của component
        > gọi `/return` nhầm trong lúc React unmount component.

        ### Kết quả
        - Ghế về `AVAILABLE`, hiển thị lại trên sơ đồ ghế cho người khác.
        - Slot hàng chờ (queue slot) được giải phóng → Worker sẽ kéo người tiếp theo vào.
        - Frontend nên điều hướng về `/event/:id/queue` sau khi gọi thành công.
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - eventId
              properties:
                eventId:
                  type: string
                  example: "a1b2c3d4-0000-0000-0000-000000000001"
                  description: ID sự kiện chứa các ghế cần trả
      responses:
        '200':
          description: Trả ghế thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Đã trả 2 ghế thành công."
                  data:
                    type: object
                    properties:
                      releasedCount:
                        type: integer
                        example: 2
                        description: Số ghế đã được nhả về AVAILABLE
        '401':
          description: Chưa đăng nhập (token không hợp lệ hoặc thiếu)
...